Requirement: #1139
Partial commit.  Further refinement of online compression support.  Changes include the following:

- Added a display of the study status on the main search screen.
- While doing the above changes, added a display of the modalities associated with a study in the study grid view.
- Added new status values for Online/Lossless compressed and Online/Lossy compressed.
- Modified the Workqueue processor to set the study process directly to Online/Lossless or Online/Lossy if we have received the object in a lossless/lossy transfer syntax.
- Made modifications to the TransferSyntax object of the tool kit, we now store bools with each transfer syntax telling if its a lossless or lossy transfer syntax.
- Ported over some changes from the 0.95 branch, specifically, some changes related to how we cleanup WorkQueue entries to make this more consistent.
- Modified the EntityBroker class so that when we do updates to a table, we don't need to have the primary key of a row, we can instead use select criteria to specify the rows to update.
- The "Specifications" sub-namespace in the ImageServer.Rules namespace was incorrectly in a wrong folder.  Moved these into the Rules/Specifications sub-namespace.
- Fixed bug with how we were setting the derivation description when doing a compression/decompression.
- Ported over some stored procedure updates from the 0.95 release related to disk management.
- Added a bool when executing the rules engine so that we can now spec if we want to stop applying rules of a specific type after we find one match of the rule.

+++ ImageServer/Services/WorkQueue/StudyProcess/StudyProcessItemProcessor.cs
@@ -359,9 +359,12 @@ namespace ClearCanvas.ImageServer.Services.WorkQueue.StudyProcess
                             string newPath = basePath + "." + extension;
                             if (!File.Exists(newPath))
                             {
-                                sop.Extension = extension;
-                                File.Move(path, newPath);
-                                break;
+								if (File.Exists(path))
+								{
+									sop.Extension = extension;
+									File.Move(path, newPath);
+								}
+                            	break;
                             }
                         }
 
@@ -488,15 +491,17 @@ namespace ClearCanvas.ImageServer.Services.WorkQueue.StudyProcess
             Console.WriteLine("WorkQueue Item is being processed...");
 #endif
             bool successful = false;
+        	bool batchProcessed = true;
             _statistics.TotalProcessTime.Add(
-                    delegate()
-                    {
+                    delegate
+                    	{
                             //Load the specific UIDs that need to be processed.
                             LoadUids(item);
 
                             if (WorkQueueUidList.Count == 0)
                             {
                                 successful = true;
+                            	batchProcessed = false;
                             }
                             else
                             {
@@ -528,13 +533,15 @@ namespace ClearCanvas.ImageServer.Services.WorkQueue.StudyProcess
 				}
 				if (failNow)
 				{
-					Platform.Log(LogLevel.Error, "Failing queue entry, all entries are duplicates.");
-					FailQueueItem(item);
+					Platform.Log(LogLevel.Error,"Failing queue entry, all entries are duplicates.");
+					PostProcessingFailure(item, true);
 					return;
 				}
 			}
-
-            PostProcessing(item, WorkQueueUidList.Count, !successful);
+			if (successful)
+				PostProcessing(item, batchProcessed, false);
+			else
+				PostProcessingFailure(item, false);
         }
